% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constructor-Modules.R
\name{AddMinFrac}
\alias{AddMinFrac}
\title{Constructor function to calculate minimum fractions across samples}
\usage{
AddMinFrac(from.table, to.table)
}
\arguments{
\item{from.table}{from which table should LUMA pull the Peak.list}

\item{to.table}{to which should LUMA save the modified Peak.list}
}
\description{
Modifies the Peak.list by calculating the minimum number of samples within a treatment class to contain a given feature.
See calc_minfrac for more details
}
