% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/In_House_Library-Modules.R
\name{AnnotatePeaklist}
\alias{AnnotatePeaklist}
\title{Annotates Peak.list with In House Library}
\usage{
AnnotatePeaklist(from.table, to.table, lib.db, ...)
}
\arguments{
\item{from.table}{from which table should LUMA pull the Peak.list}

\item{to.table}{to which should LUMA save the modified Peak.list}

\item{lib.db}{character name of In House Library database. Default is
'Annotated Library'}

\item{...}{arguments to pass to match_Annotation.}
}
\description{
Compares isotope and adduct annotations within Peaklist to
  user-defined annotation library. For more details, see
  \code{match_Annotation}. For working examples, see \code{InitWorkflow,
  AnnotatePeaklist, CombineFeatures, CombinePeaklists, CullBackground,
  CullMF, CullCV, CullVoidVolume, FormatForMetaboAnalystR, FormatForSIMCA,
  NormalizePeaklists, ParseCAMERA, SimplyPeaklists, FinalWorkflow}.
}
