% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation-functions.R
\name{calc_cv}
\alias{calc_cv}
\title{Calculate coeffient of variation (CV) by pooled QCs}
\usage{
calc_cv(Peak.list)
}
\arguments{
\item{Peak.list}{a table of class 'tbl_df',tbl' or 'data.frame' with variables as columns.  Should contain all output columns from XCMS and CAMERA.}
}
\description{
Calculates the CV value across all pooled QC samples
}
\examples{
library(LUMA)
test <- calc_cv(Peak.list = Peaklist_Pos$From_CAMERA)
test[["\%CV"]][11:23]
}
