% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation-functions.R
\name{calc_minfrac}
\alias{calc_minfrac}
\title{Calculate minimum fraction of features across sample classes}
\usage{
calc_minfrac(Sample.df, xset4, BLANK, Peak.list)
}
\arguments{
\item{Sample.df}{a data frame with class info as columns.  Must contain a separate row entry for each unique sex/class combination. Must contain the columns 'Sex','Class','n','Endogenous'.}

\item{xset4}{an xcms object that has had peak picking, retention time alignment, peak grouping, and imputing missing values performed}

\item{BLANK}{a logical indicating whether blanks are being evaluated}

\item{Peak.list}{a table of class 'tbl_df',tbl' or 'data.frame' with variables as columns.  Should contain all output columns from XCMS and CAMERA, and additional columns from IHL.search.}
}
\value{
data frame containing the original table with one additional column 'Minfrac' at the end, followed by the CAMERA columns 'isotopes','adduct','pcgroup'
}
\description{
Calculates the fraction of samples within each class that contains a given feature.
Only the minimum fraction across all sample classes is returned.
}
\examples{
  library(LUMA)
  if(require(lcmsfishdata, quietly = TRUE)) {
  file <- system.file('extdata','XCMS_objects_Pos.Rdata', package = "lcmsfishdata")
  load(file)
  file2 <- system.file('extdata','Sample_Class.txt', package = "LUMA")
  Sample.df <- read.table(file2, sep = "\t", header = TRUE) #Ignore Warning message
  Peak.list <- lcmsfishdata::Peaklist_Pos$From_CAMERA
  test <- calc_minfrac(Sample.df = Sample.df, xset4 = xset4, BLANK = FALSE,
  Peak.list = Peak.list)
  test[["MinFrac"]][11:23]
  }
}
