% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-functions.R
\name{calc_ranges}
\alias{calc_ranges}
\title{Calculates search ranges}
\usage{
calc_ranges(search.list, search.par, mycols)
}
\arguments{
\item{search.list}{tbl containing the list of mz and rt values to search against the IHL. Must contain only three columns, "EIC_ID", "mz", and "rt"}

\item{search.par}{a single-row data frame with 11 variables containing user-defined search parameters. Must contain the columns 'ppm','rt','Voidrt','Corr.stat.pos','Corr.stat.neg','CV','Minfrac','Endogenous','Solvent','gen.plots','keep.singletons'.}

\item{mycols}{a character vector containing all the names to include in the search output}
}
\value{
NULL testing
}
\description{
Calculates the min and max ranges for searching against the In House Library (IHL)
}
