% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialog-functions.R
\name{CAMERAFiles_dlg}
\alias{CAMERAFiles_dlg}
\title{CAMERA Files Dialog Box}
\usage{
CAMERAFiles_dlg(multiple = FALSE)
}
\arguments{
\item{multiple}{should dialog box with multiple fields be used; caution this is buggy and doesn't always work on all platforms; default if FALSE}
}
\value{
named list
}
\description{
Creates dialog box for user to select saved CAMERA objects when running LUMA workflow
}
\examples{
\dontrun{
library(LUMA)
if(require(lcmsfishdata, quietly = TRUE)) {
  (mypath <- system.file("extdata", package = "lcmsfishdata"))
  (test <- CAMERAFiles_dlg(multiple = TRUE))
  }
}
}
