% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser-functions.R
\name{combine_phenodata}
\alias{combine_phenodata}
\title{Combine phenotype data in Peak.list}
\usage{
combine_phenodata(Sample.df, Peak.list, Summed.list, search.par, BLANK,
  IonMode)
}
\arguments{
\item{Sample.df}{a data frame with class info as columns.  Must contain a separate row entry for each unique sex/class combination. Must contain the columns 'Sex','Class','n','Endogenous'.}

\item{Peak.list}{data frame. Must have Correlation.stat, metabolite_group, and mono_mass columns.  Should contain output columns from XCMS and CAMERA, and additional columns from IHL.search, Calc.MinFrac, CAMERA.parser and EIC.plotter functions.}

\item{Summed.list}{data frame containing metabolite group as first column and the rest summed intensities for each sample}

\item{search.par}{a single-row data frame with 11 variables containing user-defined search parameters. Must contain the columns 'ppm','rt','Voidrt','Corr.stat.pos','Corr.stat.neg','CV','Minfrac','Endogenous','Solvent','gen.plots','keep.singletons'.}

\item{BLANK}{a logical indicating whether blanks are being evaluated}

\item{IonMode}{a character string defining the ionization mode.  Must be either 'Positive' or 'Negative'}
}
\value{
Peak.list.summed with the munged phenotype columns up front followed by QC and sample columns
}
\description{
Combine phenotype data for each metabolite group in Peak.list with summed intensity values
}
