% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combine_Peaklists-Modules.R
\name{CombinePeaklists}
\alias{CombinePeaklists}
\title{Combine Peaklists from different ion modes}
\usage{
CombinePeaklists(
  from.tables,
  to.table,
  method,
  peak.db,
  db.dir,
  gen.plots,
  CAMERA.pos,
  CAMERA.neg
)
}
\arguments{
\item{from.tables}{character vector of table names to draw from databases to be combined.
If gen.plots = TRUE, should be of length 2. Otherwise, only the first element will be used.}

\item{to.table}{to which table should LUMA save the modified Peak.list}

\item{method}{which method to use for removing ion mode duplicates.
See search_IonDup for available methods.}

\item{peak.db}{what database contains the Peaklists to be combined.
Default is 'Peaklist_db'}

\item{db.dir}{directory containing the database.
Default is 'db'}

\item{gen.plots}{logical indicating whether LUMA needs to generate plots to inspect ion mode duplicates.
Default is to check whether this variable was assigned by a call to InitWorkflow}

\item{CAMERA.pos}{which CAMERA object to use to plot metabolite groups for positive mode.
Default is to read from saved R objects by call to InitWorkflow}

\item{CAMERA.neg}{which CAMERA object to use to plot metabolite groups for negative mode.
Default is to read from saved R objects by call to InitWorkflow}
}
\description{
Combines Peaklists from positive and negative ionization modes while removing ion mode duplicates.
See combine_ion_modes and search_IonDup for more details.
}
