% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-functions.R
\name{connect_libdb}
\alias{connect_libdb}
\title{Connects to library database}
\usage{
connect_libdb(lib.db, db.dir)
}
\arguments{
\item{lib.db}{character name of database}

\item{db.dir}{character directory containing the database}
}
\value{
Formal class SQLiteConnection
}
\description{
Establishes a connection to an RSQLite database for searching Peak.list against library; if doesn't exist, creates new database
}
