% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-functions.R
\name{connect_lumadb}
\alias{connect_lumadb}
\title{Connects to LUMA database}
\usage{
connect_lumadb(db.list, db.dir, new.db)
}
\arguments{
\item{db.list}{list chracter names of databases containing results from processing positive mode (1,3) and negative mode (2,4) data for samples (1,2) and blanks (3,4)}

\item{db.dir}{character directory containing the databases}

\item{new.db}{character what should the new database be called.}
}
\value{
list of Formal class SQLiteConnections, starting with new.db entry followed by one for each db.list entry
}
\description{
Establishes a connection to an RSQLite database for combining Peak.lists together from two different ionization modes.
Must have previously saved SQLite databases to hard disk.
}
