% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-functions.R
\name{connect_peakdb}
\alias{connect_peakdb}
\title{Connects to peak database}
\usage{
connect_peakdb(file.base, db.dir, mem)
}
\arguments{
\item{file.base}{character return from gen_filebase function}

\item{db.dir}{character what should the database directory be called.  Default is 'db'}

\item{mem}{logical should database be in-memory. Default is FALSE}
}
\value{
Formal class SQLiteConnection
}
\description{
Establishes a connection to an SQLite database for storing Peak.list; if doesn't exist, creates new database
}
