% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Culling-Modules.R
\name{CullCV}
\alias{CullCV}
\title{Culls the Peak.list by Percent CV}
\usage{
CullCV(from.table, to.table)
}
\arguments{
\item{from.table}{from which table should LUMA pull the Peak.list}

\item{to.table}{to which should LUMA save the modified Peak.list}
}
\description{
Culls the Peak.list by removing all components with a
  coefficient of variation (CV) greater than the user-specified cutoff across
  all Pooled QC samples. See \code{trim_cv} for more details.
}
\examples{
\dontrun{
library(LUMA)
db.dir <- system.file('extdata/', package = "LUMA")
InitWorkflow(db.dir = db.dir)
CullCV(from.table = "From CAMERA", to.table = "Trimmed by CV")
}
}
