% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Culling-Modules.R
\name{CullVoidVolume}
\alias{CullVoidVolume}
\title{Culls the Peak.list by Void volume}
\usage{
CullVoidVolume(from.table, to.table, method)
}
\arguments{
\item{from.table}{from which table should LUMA pull the Peak.list}

\item{to.table}{to which should LUMA save the modified Peak.list}

\item{method}{which method to apply to trim by retention time.  See
\code{trim_rt} for details.}
}
\description{
Culls the Peak.list by removing all components coming from the
  void volume. See \code{remove_void_volume()} for more details. For
  examples, see \code{InitWorkflow()} and vignettes.
}
