% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{EIC.plotter}
\alias{EIC.plotter}
\title{Loop-based EIC Plotter}
\usage{
EIC.plotter(anposGa, Sample.df, Peak.list, center, BLANK, gen.plots, ion.mode,
  file.base, QC.id)
}
\arguments{
\item{anposGa}{xsannotate object with annotated isotopes and ion adducts and fragments}

\item{Sample.df}{a data frame with class info as columns.  Must contain a separate row entry for each unique sex/class combination. Must contain the columns "Sex","Class","n","Endogenous".}

\item{Peak.list}{a data frame from CAMERA that has been parsed.  Should contain all output columns from XCMS and CAMERA, and additional columns from IHL.search, Calc.MinFrac and CAMERA.parser.}

\item{center}{numeric value indicating which sample to pick for plotting purposes}

\item{BLANK}{a logical indicating whether blanks are being evaluated. Default is FALSE}

\item{gen.plots}{a logical indicating whether to create plots for metabolite groups.  Default is FALSE}

\item{ion.mode}{a character string defining the ionization mode.  Must be either "Positive" or "Negative"}

\item{file.base}{character string used to name graphical output}

\item{QC.id}{character identifier for pooled QC samples. Default is "Pooled_QC"}
}
\value{
List of length 2.  1st element is a data frame with all columns as the original data frame with one additional column "Correlation.stat".  2nd element is a data frame specifically used to validate CAMERA results.
}
\description{
Plots EICs and Pseudospectra from (parsed) metabolite groups using plotEICs-methods and plotPsSpectrum-methods from CAMERA. Also plots correlation matrices and clustered dendrograms for each (parsed) metabolite group.
}
