% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting-functions.R
\name{format_simca}
\alias{format_simca}
\title{Formats Peak.list for SIMCA}
\usage{
format_simca(
  Peak.list = NULL,
  Sample.df,
  Sample.data,
  tbl.id = NULL,
  QC.id = "Pooled_QC_",
  ...
)
}
\arguments{
\item{Peak.list}{data frame containing combined ion mode peaklist with ion mode duplicates removed.  Alternatively can be retrieved from databases.  Default is NULL}

\item{Sample.df}{data frame with class info as columns.  Must contain a separate row entry for each unique sex/class combination. Must contain the columns 'Sex','Class','n','Endogenous'.}

\item{Sample.data}{data frame with phenotype data as columns and a row for each study sample.  First column must be a unique sample identifier with the header 'CT-ID'.  Phenotype columns may vary, but must include two columns called 'Plate Number' and 'Plate Position' for determining run order.
'Plate Number' must be numeric and is equivalent to batch number.  'Plate Position' must be alphanumeric and corresponds to row(alpha) and column(numeric) positions on, e.g. a 96-well plate}

\item{tbl.id}{character Table name to draw from database. Default is NULL}

\item{QC.id}{character vector specifying identifier in filename designating a Pooled QC sample.  Only the first value will be used.  Default is 'Pooled_QC_'}

\item{...}{Arguments to pass parameters to database functions}
}
\value{
NULL testing
}
\description{
Formats Peak.list for import into SIMCA
}
\examples{
library(LUMA)
if(require(lcmsfishdata, quietly = TRUE)) {
  # is case sensitive on Linux
  file <- system.file('extdata','Sample_Class.txt', package = "lcmsfishdata")
  Sample.df <- read.table(file, header = TRUE, sep = "\t")
 # is case sensitive on Linux
 file2 <- system.file('extdata','Sample_Data.csv', package = "lcmsfishdata")
  Sample.data <- read.table(file2, header = TRUE, sep = ",")
 Peak.list <- Peaklist_db$Peaklist_Normalized
  test <- format_simca(Peak.list = Peak.list, Sample.df = Sample.df, Sample.data = Sample.data)
}
}
