% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting-functions.R
\name{format_stats}
\alias{format_stats}
\title{Formats for statistical analysis}
\usage{
format_stats(Peak.list, Sample.df)
}
\arguments{
\item{Peak.list}{Data table from output of any LUMA function}

\item{Sample.df}{data frame with class info as columns.  Must contain a separate row entry for each unique sex/class combination for statistical analysis. Must contain the columns 'Sex','Class','n','Endogenous'.}
}
\value{
list of length 2. 1st element is a dataframe of (mostly) numeric values from a metabolomics dataset with samples (rows) and metabolite intensities (cols).
1st column contains a factor vector describing the sample class.
2nd element in returned list is a character string of the endogenous (control) class
}
\description{
Formats LUMA output for use with statistical functions
}
