% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_Output-Modules.R
\name{FormatForMetaboAnalystR}
\alias{FormatForMetaboAnalystR}
\title{Formatting of metabolite data for MetaboAnalystR.}
\usage{
FormatForMetaboAnalystR(
  from.table,
  to.csv,
  data.type = "pktable",
  anal.type = "stat",
  paired = FALSE,
  peak.db,
  db.dir
)
}
\arguments{
\item{from.table}{from which table should LUMA read the Peaklist}

\item{to.csv}{to what filename (excluding .csv extension) should LUMA save the formatted Peaklist}

\item{data.type}{What type of data will be generated. See usage and format_MetabolomicData for options.}

\item{anal.type}{character Indicates the analysis module the data will be used for. See usage and documentation for MetaboAnalystR::InitDataObjects() for options.}

\item{paired}{logical Indicate if the data is paired or not.
Default is FALSE}

\item{peak.db}{what database contains the Peaklists to be combined.
Default is 'Peaklist_db'}

\item{db.dir}{directory containing the database.
Default is 'db'}
}
\value{
mSetObj
}
\description{
This function initializes objects that will hold the metabolite data, formats peak intensity data into one of the formats acceptable by MetaboAnalystR, and sets the metabolite data object.
}
\examples{
\dontrun{
library(LUMA)
db.dir <- system.file('extdata/', package = "LUMA")
InitWorkflow(db.dir = db.dir)
AnnotatePeaklist(from.table = "From CAMERA", to.table = "Annotated")
FormatForMetaboAnalystR(from.table = "Annotated", to.csv = "Peaklist_for_MetaboAnalyist", peak.db = peak_db, db.dir = db.dir)
}
}
