% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-functions.R
\name{gen_adducts}
\alias{gen_adducts}
\title{Generates adducts for in house library}
\usage{
gen_adducts(Annotated.library, rules, ion.mode)
}
\arguments{
\item{Annotated.library}{a data frame with annotated metabolite entries. Must contain columns called 'Name', 'Formula', Molecular.Weight' and 'RT..Min.'.  Can contain additional info as separate columns.}

\item{rules}{a data frame containing the rule list used by CAMERA to annotate ion adducts and fragments.  Must contain the columns 'name','nmol','charge','massdiff','oidscore','quasi','ips'.}

\item{ion.mode}{a character string defining the ionization mode.  Must be either 'Positive' or 'Negative'}
}
\value{
NULL testing
}
\description{
Generates a list of adducts for all metabolites in the supplied annotation library based on the rules input
}
