% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-functions.R
\name{get_features}
\alias{get_features}
\title{Retrieves features from database}
\usage{
get_features(myname, peak.db, asdf)
}
\arguments{
\item{myname}{character name of table in database to return}

\item{peak.db}{Formal class SQLiteConnection}

\item{asdf}{logical indicating whether to return a data frame instead of a tibble. Default is FALSE}
}
\value{
tbl alternatively a data frame
}
\description{
Returns mz/rt features from a RSQLite database
}
