% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical-functions.R
\name{get_pvalue}
\alias{get_pvalue}
\title{get_pvalue}
\usage{
get_pvalue(values, classes, stat.method)
}
\arguments{
\item{values}{a dataframe of numeric values from a metabolomics dataset with samples(rows) and metabolite intensity values (cols)}

\item{classes}{a vector of factor values describing the sample class}

\item{stat.method}{which statistical method to use. Can be ANOVA, ANOVA_FDR, or Ttest}
}
\value{
a matrix containing p-values, F-stats, degrees of freedom for class and
degrees of freedom for residuals for each metabolite (row)
}
\description{
Get p-value and other statistical parameters from univariate statistical analysis of metabolomics dataset
}
\examples{
get_pvalue(metabolites, grouping, "ANOVA")
get_pvalue(metabolites, grouping, "Ttest")

subset <- which(grouping == "100_Effluent"|grouping == "Control")
mydata <- metabolites[subset,]
myclass <- grouping[subset]
get_pvalue(mydata,myclass,"Ttest")
}
