% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{grouping}
\alias{grouping}
\title{Factors for the samples (rows) in the values dataset}
\format{A factor-formatted vector with 48 values
\describe{
\item{classes}{grouping factor corresponding to the sample class info for liver samples in the values dataset}
}}
\source{
Unpublished data from the EPA metabolomics team in Athens, GA
}
\usage{
grouping
}
\description{
A vector containing factor values to use for statistical analysis for the 48 liver samples in the values dataset
}
\keyword{datasets}
