% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialog-functions.R
\name{InputFiles_dlg}
\alias{InputFiles_dlg}
\title{Input Files Dialog Box}
\usage{
InputFiles_dlg(WorkingDir, multiple = FALSE)
}
\arguments{
\item{WorkingDir}{What is the working directory for the LUMA workflow}

\item{multiple}{should dialog box with multiple fields be used; caution this is buggy and doesn't always work on all platforms; default if FALSE}
}
\value{
named list
}
\description{
Creates dialog box for user to select input files when initiating LUMA workflow
}
\examples{
\dontrun{
library(LUMA)
if(require(lcmsfishdata, quietly = TRUE)) {
  (mypath <- system.file("extdata", package = "lcmsfishdata"))
  (test <- InputFiles_dlg(WorkingDir = mypath, multiple = TRUE))
  }
}
}
