% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-functions.R
\name{is.whole}
\alias{is.whole}
\title{Is whole number}
\usage{
is.whole(a)
}
\arguments{
\item{a}{a vector or array to test}
}
\value{
logical
}
\description{
Tests if a vector or array contains only whole numbers
}
