% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-functions.R
\name{libdbConnect}
\alias{libdbConnect}
\title{libdbConnect}
\usage{
libdbConnect(lib.db, db.dir)
}
\arguments{
\item{lib.db}{character name of database}

\item{db.dir}{character directory containing the database}
}
\value{
Formal class SQLiteConnection
}
\description{
Establishes a connection to an RSQLite database for library searching; if doesn't exist, creates new database
}
