% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-functions.R
\name{LUMA_dbConnect}
\alias{LUMA_dbConnect}
\title{LUMA_dbConnect}
\usage{
LUMA_dbConnect(db.list, db.dir, new.db)
}
\arguments{
\item{db.list}{list chracter names of databases containing results from processing positive mode (1,3) and negative mode (2,4) data for samples (1,2) and blanks (3,4)}

\item{db.dir}{character directory containing the databases}

\item{new.db}{character what should the new database be called}
}
\value{
list of Formal class SQLiteConnections, starting with new.db entry followed by one for each db.list entry and
}
\description{
Establishes a connection to an RSQLite database for combining two datasets together from two different ionization modes
}
