% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mSetObj.R
\docType{data}
\name{mSetObj}
\alias{mSetObj}
\title{Example mSetObj}
\format{
nested list of class \code{pktable} \describe{
  \item{dataSet}{A list with elements \code{type}, \code{design.type}, \code{cls.type}, and \code{format}}
  \item{analSet}{An empty list}
  \item{imgSet}{An empty list}
  \item{msgSet}{A list with the element \code{msg.vec}}
  \item{cmdSet}{an empty character vector}
   }
}
\source{
https://https://github.com/jmosl01/LUMA/blob/develop/data-raw/mSetObj.R
}
\usage{
mSetObj
}
\description{
object of class mSetObj generated from scratch in LUMA
}
\examples{
class(mSetObj)
}
\keyword{datasets}
