% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_Zeros_and_Normalize-Algorithm.R
\name{normalize_unit}
\alias{normalize_unit}
\title{normalize_unit}
\usage{
normalize_unit(x)
}
\arguments{
\item{x}{dataframe containing only numeric vectors representing samples with metabolite intensities.  Cannot contain zeros or NA values}
}
\value{
An array with the same shape as \code{x}, but with the summary statistics swept out
}
\description{
normalizes each column to unity
}
