% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting-functions.R
\name{output_MetaData}
\alias{output_MetaData}
\alias{output_MetaData.pktable}
\alias{output_MetaData.mass_all}
\alias{output_MetaData.default}
\title{Printing metadata as CSV files for MetaboAnalystR.}
\usage{
output_MetaData(mSetObj, Peak.list, Sample.df, Sample.data)

\method{output_MetaData}{pktable}(mSetObj, Peak.list, Sample.df, Sample.data)

\method{output_MetaData}{mass_all}(mSetObj, Peak.list, Sample.df, Sample.data)

\method{output_MetaData}{default}(mSetObj, Peak.list, Sample.df, Sample.data)
}
\arguments{
\item{mSetObj}{NULL}

\item{Peak.list}{data frame containing combined ion mode peaklist with ion mode duplicates removed.}

\item{Sample.df}{data frame with class info as columns.  Must contain a separate row entry for each unique sex/class combination. Must contain the columns 'Sex','Class','n','Endogenous'.}

\item{Sample.data}{data frame with phenotype data as columns and a row for each study sample.  First column must be a unique sample identifier with the header 'CT-ID'.  Phenotype columns may vary, but must include two columns called 'Plate Number' and 'Plate Position' for determining run order.}
}
\value{
mSetObj
}
\description{
This function prints the metabolite and sample metadata to CSV files.
}
\examples{
\dontrun{
library(LUMA)
file <- system.file("extdata","Sample_Class.txt", package = "LUMA") # is case sensitive on Linux
Sample.df <- read.table(file, header = TRUE, sep = "\t")
file2 <- system.file("extdata","Sample_Data.csv", package = "LUMA") # is case sensitive on Linux
Sample.data <- read.table(file2, header = TRUE, sep = ",")
Peak.list <- Peaklist_db$Peaklist_Normalized
mSetObj <- NULL
output_MetaData(mSetObj = mSetObj, Peak.list =
Peak.list, Sample.df = Sample.df, Sample.data = Sample.data)
}
}
