% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Munge_Data-Modules.R
\name{ParseCAMERA}
\alias{ParseCAMERA}
\title{Parses CAMERA annotations}
\usage{
ParseCAMERA(from.table, to.table)
}
\arguments{
\item{to.table}{to which should LUMA save the modified Peak.list}
}
\description{
Parses the three annotation columns generated by CAMERA to remove redundant and duplicative ion adduct annotations
See parse_pos_results, parse_neg_results and plot_metgroup for more details.
#' @param from.table from which table should LUMA pull the Peak.list
}
