% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-functions.R
\name{peakdbConnect}
\alias{peakdbConnect}
\title{peakdbConnect}
\usage{
peakdbConnect(file.base, db.dir)
}
\arguments{
\item{file.base}{character return from filebase.gen function}

\item{db.dir}{character what should the database directory be called.  Default is "db"}
}
\value{
Formal class SQLiteConnection
}
\description{
Establishes a connection to an RSQLite database for storing data; if doesn't exist, creates new database
}
