% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ion-mode-functions.R
\name{pre_combine_ion_modes}
\alias{pre_combine_ion_modes}
\title{Combines Peak.lists Simply}
\usage{
pre_combine_ion_modes(Peak.list = NULL, tbl.id, ...)
}
\arguments{
\item{Peak.list}{a named list of data frames (one per ionization mode)
containing intensity matrices across all study samples and Pooled QCs.
Names must be \code{c("Positive","Negative")}.  Alternatively may use existing
database connections by setting to NULL and specifying database parameters
with ...}

\item{tbl.id}{character vector of table names to draw from database.  First
value should be table name for positive mode, second should be table name
for negative mode. Default is NULL}

\item{...}{arguments to pass to database functions}
}
\value{
NULL testing
}
\description{
Combines Peak.lists from both ionization modes (positive and
  negative) for a first look at class separations, e.g. with PCA.
}
\examples{
library(LUMA)
if(require(lcmsfishdata, quietly = TRUE)) {

  Peak.list <- list(Positive = lcmsfishdata::Peaklist_Pos[["From_CAMERA"]],
                    Negative = lcmsfishdata::Peaklist_Neg[["From_CAMERA"]])
  test <- pre_combine_ion_modes(Peak.list = Peak.list)
  nrow(test) #Combined features from simply combining Peaklists
  sum(sapply(Peak.list, nrow)) #Doesn't remove any ion mode duplicates
}
}
