% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Exogenous_and_Solvent_Peak-Algorithm.R
\name{PreCombineIonModes}
\alias{PreCombineIonModes}
\title{PreCombineIonModes}
\usage{
PreCombineIonModes(Peak.list, tbl.id, ...)
}
\arguments{
\item{Peak.list}{a named list of data frames \(one per ionization mode\) containing intensity matrices across all study samples and Pooled QCs.  Names must be c("Positive","Negative").  Alternatively may use existing database connections by setting to NULL and specifying database parameters with ...}

\item{tbl.id}{character vector of table names to draw from database.  First value should be table name for positive mode, second should be table name for negative mode. Default is NULL}

\item{...}{Arguments to pass parameters to database functions}
}
\value{
NULL testing
}
\description{
Combines peaklists from both ionization modes \(positive and negative\) for a first look at class separations
}
