% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-functions.R
\name{read_tbl}
\alias{read_tbl}
\title{Reads Peak.list from database}
\usage{
read_tbl(mytable, peak.db, asdf)
}
\arguments{
\item{mytable}{character name of table in database to return}

\item{peak.db}{Formal class SQLiteConnection}

\item{asdf}{logical indicating whether to return a data frame instead of a tibble. Default is FALSE}
}
\value{
tbl alternatively a data frame
}
\description{
Extracts Peak.list from an RSQLite database as a tibble.
Alternatively, load the Peak.list into memory as a data frame
}
