% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-functions.R
\name{Readtbl}
\alias{Readtbl}
\title{Readtbl}
\usage{
Readtbl(myname, peak.db, asdf)
}
\arguments{
\item{myname}{character name of table in database to return}

\item{peak.db}{Formal class SQLiteConnection}

\item{asdf}{logical indicating whether to return a data frame instead of a tibble. Default is FALSE}
}
\value{
tbl alternatively a data frame
}
\description{
Extract table from an RSQLite database as a tibble.  Alternatively load into memory as a data frame
}
