% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimming-functions.R
\name{remove_void_volume}
\alias{remove_void_volume}
\title{Removes void volume from Peak.list}
\usage{
remove_void_volume(Peak.list, search.par, method, ...)
}
\arguments{
\item{Peak.list}{data frame. Must have Correlation.stat column.  Should
contain output columns from XCMS and CAMERA, and additional columns from
IHL.search, Calc.MinFrac, Calc.corr.stat and EIC.plotter functions.}

\item{search.par}{a single-row data frame with 11 variables containing
user-defined search parameters. Must contain the columns
\code{"ppm","rt","Voidrt","Corr.stat.pos","Corr.stat.neg","CV","Minfrac","Endogenous","Solvent","gen.plots","keep.singletons"}.}

\item{method}{which method to apply to trim by retention time.  See trim_rt
for details}

\item{...}{Arguments to pass to trim_rt}
}
\value{
data frame Peak.list.trimmed original Peak.list without all
metabolite groups containing at least one feature in the void volume
}
\description{
Removes features or compounds found in the void volume of the
chromatographic run from the Peak.list
}
\examples{
library(LUMA)
if(require(lcmsfishdata, quietly = TRUE)){
# is case sensitive on Linux
file <- system.file('extdata','Search_Parameters.txt', package = "LUMA")
search.par <- read.table(file, sep = "\t", header = TRUE)
test <- remove_void_volume(Peak.list = lcmsfishdata::Peaklist_Pos$From_CAMERA, search.par =
search.par, method = "mz")

#Removes 275 features within the void volume
nrow(lcmsfishdata::Peaklist_Pos$From_CAMERA) - nrow(test)
}
}
