% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Replace_Zeros_and_Normalize-Algorithm.R
\name{ReplaceZerosandNormalize}
\alias{ReplaceZerosandNormalize}
\title{ReplaceZerosandNormalize}
\usage{
ReplaceZerosandNormalize(Peak.list = NULL, Sample.df, tbl.id = NULL,
  QC.id = "Pooled_QC_", mult = 1000, ...)
}
\arguments{
\item{Peak.list}{a data frame containing combined ion mode peaklist with ion mode duplicates removed.  Alternatively can be retrieved from databases.  Default is NULL}

\item{Sample.df}{a data frame with class info as columns.  Must contain a separate row entry for each unique sex/class combination. Must contain the columns "Sex","Class","n","Endogenous".}

\item{tbl.id}{character vector of table names to draw from databases.  First value should be table name from positive ionization peak database, second should be table name from negative ionization peak database. Default is NULL}

\item{QC.id}{character vector specifying identifier in filename designating a Pooled QC sample.  Only the first value will be used.  Default is "Pooled_QC_"}

\item{mult}{numeric multiplier to be added to unit normalization.  Default is 1000}

\item{...}{Arguments to pass parameters to database functions}
}
\value{
data frame containing normalized intensity matrix for all samples plus Pooled QCs
}
\description{
Replaces zero values with half of the minimum value from all study samples
}
