% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical-functions.R
\name{run_Fold_Change}
\alias{run_Fold_Change}
\title{run_Fold_Change}
\usage{
run_Fold_Change(values, classes, control)
}
\arguments{
\item{values}{a dataframe of numeric values from a metabolomics dataset with samples (rows) and metabolite intensities (cols)}

\item{classes}{a vector of factor values describing the sample class}

\item{control}{string name of the relative class against which fold change will be calculated}
}
\value{
a dataframe with first column containing descriptors for mean fold change values and standard error of the means. Contains values for each sample class (rows)
 and metabolite (cols).
}
\description{
Calculates fold change relative to controls for an untargeted metabolomics dataset
}
\examples{
run_Fold_Change(metabolites,grouping,"Control")
}
