% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mummichog-functions.R
\name{run_mummichog}
\alias{run_mummichog}
\title{Runs _mummichog_}
\usage{
run_mummichog(Sample.df, Peak.list, ion.mode, stat.method, search.par)
}
\arguments{
\item{Sample.df}{a data frame with class info as columns.  Must contain a separate row entry for each unique sex/class combination. Must contain the columns 'Sex','Class','n','Endogenous'.}

\item{Peak.list}{a table of class 'tbl_df',tbl' or 'data.frame' with variables as columns.  Should contain all output columns from XCMS and CAMERA.}

\item{ion.mode}{a character string defining the ionization mode.  Must be either 'Positive' or 'Negative'}

\item{stat.method}{statistical method to use. Can be "TTest" or "ANOVA"}

\item{search.par}{a single-row data frame with 11 variables containing user-defined search parameters. Must contain the columns 'ppm','rt','Voidrt','Corr.stat.pos','Corr.stat.neg','CV','Minfrac','Endogenous','Solvent','gen.plots','keep.singletons'.}

\item{statistic}{what statistic should be used in _mummichog_. Can be "t-score" or "Fold change"}
}
\value{
NULL testing
}
\description{
Runs _mummichog_ within the R environment.  Basically a wrapper script built from the "MS Peaks to Pathways" vignette from MetaboAnalystR.
}
