% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-functions.R
\name{search_IHL}
\alias{search_IHL}
\title{Searches Peak.list against In House Library}
\usage{
search_IHL(Peak.list, Annotated.library, rules, search.par, ion.mode,
  lib_db)
}
\arguments{
\item{Peak.list}{a table of class 'tbl_dbi', 'tbl_sql', 'tbl_lazy', or 'tbl' with samples as columns.  Should contain all output columns from XCMS and CAMERA, both metadata and sample data. Retention times must be in min.}

\item{Annotated.library}{a data frame with annotated metabolite entries. Must contain columns called 'Name', 'Formula', Molecular.Weight' and 'RT..Min.'.  Can contain additional info as separate columns.}

\item{rules}{a data frame containing the rule list used by CAMERA to annotate ion adducts and fragments.  Must contain the columns 'name','nmol','charge','massdiff','oidscore','quasi','ips'.}

\item{search.par}{a single-row data frame with 11 variables containing user-defined search parameters. Must contain the columns 'ppm','rt','Voidrt','Corr.stat.pos','Corr.stat.neg','CV','Minfrac','Endogenous','Solvent','gen.plots','keep.singletons'.}

\item{ion.mode}{a character string defining the ionization mode.  Must be either 'Positive' or 'Negative'}

\item{lib_db}{RSQLite connection}
}
\value{
data frame containing the original table with added columns 'Name','MS.ID','Formula','Annotated.adduct' and any additional info columns from Annotated.library
}
\description{
Compares CAMERA isotope annotation within Peak.list to user-defined annotation library and returns
annotation results
}
