% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-functions.R
\name{search_solv}
\alias{search_solv}
\title{Searches for background components}
\usage{
search_solv(Peak.list, Solv.list, search.par, ...)
}
\arguments{
\item{Peak.list}{data table containing sample data}

\item{Solv.list}{data table containing blank data}

\item{...}{arguments to be passed to other functions}
}
\value{
data frame containing the original table with added columns 'Name','MS.ID','Formula','Annotated.adduct' and any additional info columns from Annotated.library
}
\description{
Searches data table from samples for components coming from background
}
