% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialog-functions.R
\name{select_rt}
\alias{select_rt}
\title{Selects retention time column}
\usage{
select_rt(annotated_library)
}
\arguments{
\item{annotated_library}{a data frame with annotated metabolite entries. Must contain a column for each of `name`, `formula`, `molecular weight`, and `retention time`.  Can contain additional info as separate columns.}
}
\value{
`character` multiple selections from column names in `annotated_library`
}
\description{
Allows the user to select the retention time column from their Annotated Library
}
