% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combine_Peaklists-Modules.R
\name{SimplyPeaklists}
\alias{SimplyPeaklists}
\title{Combine Peaklists simply}
\usage{
SimplyPeaklists(from.tables, peak.db, db.dir)
}
\arguments{
\item{from.tables}{character vector of table names to draw from databases to be combined simply.}

\item{peak.db}{what database contains the Peaklists to be combined simply.
Default is 'Peaklist_db'}

\item{db.dir}{directory containing the database.
Default is 'db'}
}
\value{
the file 'Peaklist_Combined.csv" is written to the working directory
}
\description{
Combines Peaklists from positive and negative ionization modes without removing ion mode duplicates.
See
}
