% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constructor-Modules.R
\name{start_LUMAWorkflow}
\alias{start_LUMAWorkflow}
\title{Constructor function to initiate LUMA Workflow}
\usage{
start_LUMAWorkflow(ion.id, blanks.dir, db.dir, adduct.files, use.CAMERA,
  use.XCMS, graph.method)
}
\arguments{
\item{ion.id}{character vector specifying identifier in mz data filenames designating positive or negative ionization or both.
Default is c('Pos','Neg')}

\item{blanks.dir}{character name of subdirectory containing process blank data files.
Default is 'Blanks'}

\item{db.dir}{character name of subdirectory to store databases
Default is 'db'
Positive identifier must come first. Default is c('Pos','Neg')}

\item{adduct.files}{character vector specifying file names of csv files containing adduct rules, whose file structure is inherited from CAMERA.
Positive mode adducts must come first. Default is c("primary_adducts_pos.csv","primary_adducts_neg.csv")}

\item{use.CAMERA}{logical indicating whether to run CAMERA.
Default is to look for saved CAMERA objects and run CAMERA if missing.}

\item{use.XCMS}{logical indicating whether to run XCMS.
Default is to look for saved XCMS objects and run XCMS if missing.}

\item{graph.method}{graphing method to use for CAMERA.
Default is 'lpc'. See CAMERA documentation for details.}
}
\value{
global variables are returned
}
\description{
Creates the first Peak.list and sets up storing and passing Peak.lists and ancillary data between modules.
}
