% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimming-functions.R
\name{Trim.RT}
\alias{Trim.RT}
\title{Trim by void volume}
\usage{
Trim.RT(Peak.list, search.par)
}
\arguments{
\item{Peak.list}{data frame. Must have Correlation.stat column.  Should contain output columns from XCMS and CAMERA, and additional columns from IHL.search, Calc.MinFrac, Calc.corr.stat and EIC.plotter functions.}

\item{search.par}{a single-row data frame with 11 variables containing user-defined search parameters. Must contain the columns 'ppm','rt','Voidrt','Corr.stat.pos','Corr.stat.neg','CV','Minfrac','Endogenous','Solvent','gen.plots','keep.singletons'.}
}
\value{
data frame Peak.list.trimmed original Peak.list without all metabolite groups containing at least one feature in the void volume
}
\description{
Removes features found in the void volume of the chromatographic run
}
