% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-functions.R
\name{write_tbl}
\alias{write_tbl}
\title{Writes Peak.list to database}
\usage{
write_tbl(mydf, peak.db, myname)
}
\arguments{
\item{mydf}{tbl or dataframe to write}

\item{peak.db}{Formal class SQLiteConnection}

\item{myname}{character what should the table be called}
}
\value{
a tbl object in the remote source
}
\description{
Writes Peak.list to an RSQLite database. Peak.list can be a tibble or data frame.
}
