% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{write_xlsx}
\alias{write_xlsx}
\title{Writes xlsx table}
\usage{
write_xlsx(validate.sheets, file.base, myname, mysheets)
}
\arguments{
\item{validate.sheets}{list of sheets to write to xlsx file.  Currently must be of length 2}

\item{file.base}{character return from gen_filebase function}

\item{myname}{name to append to file.base to create file name for xlsx}

\item{mysheets}{character vector to name sheets in xlsx file. Currently must be of length 2}
}
\value{
class jobjRef object
}
\description{
Write xlsx table output from validate_metgroup.  Essentially a wrapper for xlsx::saveWorkbook
}
