% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{xcmsWrap}
\alias{xcmsWrap}
\title{XCMS Wrapper}
\usage{
xcmsWrap(mzdatafiles, XCMS.par)
}
\arguments{
\item{mzdatafiles}{a character vector of data files with full path names}

\item{XCMS.par}{a single-row data frame with 13 variables containing XCMS parameters. The column names must be c("Peakwidth1","Peakwidth2","ppm","noise","snthresh","mzdiff","prefilter1","prefilter2","center","gapInit","bw","mzwid","minfrac")}
}
\value{
two XCMS objects xset and xset4 without and with retention time alignment, peak grouping, and imputing missing values
}
\description{
Run XCMS with user defined input parameters and return xcms objects
}
