% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculation-functions.R
\name{calc_corrstat}
\alias{calc_corrstat}
\title{Calculate correlation matrices for metabolite groups}
\usage{
calc_corrstat(Sample.df, Peak.list, get.mg, BLANK, IonMode)
}
\arguments{
\item{Sample.df}{a data frame with class info as columns.  Must contain a separate row entry for each unique sex/class combination. Must contain the columns 'Sex','Class','n','Endogenous'.}

\item{Peak.list}{a data frame from CAMERA that has been parsed.  Should contain all output columns from XCMS and CAMERA, and additional columns from IHL.search, Calc.MinFrac and CAMERA.parser.}

\item{get.mg}{numerical vector of metabolite groups that have more than one feature}

\item{BLANK}{a logical indicating whether blanks are being evaluated}

\item{IonMode}{a character string defining the ionization mode.  Must be either 'Positive' or 'Negative'}
}
\value{
Peak.list of class 'tbl_df',tbl' or 'data.frame' with variables as columns.  Has all of the columns as the original data frame with one additional column 'Correlation.stat'
}
\description{
Calculates the correlation matrices for metabolite groups based on the best feature within the group that belongs to the primary metabolite.
}
\examples{
 library(LUMA)
 library(lcmsfishdata)
 file <- system.file('extdata/CAMERA_objects_Pos.Rdata', package = "lcmsfishdata")
 load(file)
 pspec.length <- sapply(anposGa@pspectra, function(x) length(x))
 get.mg <- which(pspec.length > 1)
 file2 <- system.file('extdata/Sample_Class.txt', package = "LUMA")
 Sample.df <- read.table(file2, sep = "\t", header = TRUE) #Ignore Warning message
 test <- calc_corrstat(Sample.df = Sample.df, Peak.list =
 Peaklist_Pos_db$input_parsed, get.mg = get.mg, BLANK = FALSE, IonMode =
 "Positive")
 test[["Correlation.stat"]][11:23]
}
