% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combine_Ion_Modes-Modules.R
\name{combine_ion_modes}
\alias{combine_ion_modes}
\title{Combines Peak.lists from two ion modes}
\usage{
combine_ion_modes(Peak.list, search.par, ion.id, QC.id, tbl.id, method,
  ...)
}
\arguments{
\item{Peak.list}{a named list of data frames (one per ionization mode) containing intensity matrices across all study samples and Pooled QCs.
Names must be c('Positive','Negative').  Alternatively may use existing database connections by setting to NULL}

\item{search.par}{a single-row data frame with 11 variables containing user-defined search parameters. Must contain the columns 'ppm','rt','Voidrt','Corr.stat.pos','Corr.stat.neg','CV','Minfrac','Endogenous','Solvent','gen.plots','keep.singletons'.}

\item{ion.id}{character vector specifying identifier in filename designating positive or negative ionization mode or both.  Positive identifier must come first. Default is c('Pos','Neg')}

\item{QC.id}{character vector specifying identifier in filename designating a Pooled QC sample.  Only the first value will be used.  Default is 'Pooled_QC_'}

\item{tbl.id}{character vector of table names to draw from databases.  First value should be table containing compounds from positive ionization, second should be table containing compounds from negative ionization. Default is NULL}

\item{method}{which method to apply to search for duplicate entries.  See search_IonDup for details.}

\item{...}{Arguments to pass parameters to search_IonDup}
}
\value{
data frame containing the intensity matrix for the peaklist with Duplicate IDs
}
\description{
Combines the two single adduct ion mode feature tables into one for EIC plotting
}
