% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Munge_Data-Modules.R
\name{CombineFeatures}
\alias{CombineFeatures}
\title{Combines Features into Metabolites}
\usage{
CombineFeatures(from.table, to.table)
}
\arguments{
\item{from.table}{from which table should LUMA pull the Peaklist.}

\item{to.table}{to which should LUMA save the modified Peaklist.}
}
\description{
Combines all well-behaved features in metabolite groups into a
  single entry per metabolite. See \code{sum_features(), combine_phenodata()}
  for more details. For examples, see \code{InitWorkflow()} and vignettes.
}
