% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Munge_Data-Modules.R
\name{CombineFeatures}
\alias{CombineFeatures}
\title{Combines Features into Metabolites}
\usage{
CombineFeatures(from.table, to.table)
}
\arguments{
\item{from.table}{from which table should LUMA pull the Peak.list}

\item{to.table}{to which should LUMA save the modified Peak.list}
}
\description{
Combines all well-behaved features in metabolite groups into a single entry per metabolite.
See sum_features and combine_phenodata for more details.
}
\examples{
\dontrun{
library(LUMA)
db.dir <- system.file('extdata/', package = "LUMA")
InitWorkflow(db.dir = db.dir)
CombineFeatures(from.table = "From CAMERA", to.table = "Combined Isotopes and Adducts")
}
}
