% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-functions.R
\name{connect_libdb}
\alias{connect_libdb}
\title{Connects to library database}
\usage{
connect_libdb(lib.db, db.dir, mem)
}
\arguments{
\item{lib.db}{character name of database}

\item{db.dir}{character directory containing the database. Default is 'db'}

\item{mem}{logical should database be in-memory. Default is TRUE}
}
\value{
Formal class SQLiteConnection
}
\description{
Establishes a connection to an SQLite database for searching
  Peak.list against library; if doesn't exist, creates new database
}
\examples{
library(LUMA)
if(require(RSQLite, quietly = TRUE)) {
lib_db <- connect_libdb(mem = TRUE)
dbIsValid(lib_db) #Valid database is created
dbDisconnect(lib_db)
}
}
