% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Constructor-Modules.R
\name{FinalWorkflow}
\alias{FinalWorkflow}
\title{Constructor function to finalize LUMA workflow}
\usage{
FinalWorkflow(peak_db, lib_db)
}
\arguments{
\item{peak_db}{existing peak database connection}

\item{lib_db}{existing library database connection}
}
\description{
All LUMA workflows must end with this function.  Records the names of existing data tables and writes out the LUMA log file for traceability.
}
\examples{
\dontrun{
library(LUMA)
db.dir <- system.file('extdata/', package = "LUMA")
InitWorkflow(db.dir = db.dir)
AnnotatePeaklist(from.table = "From CAMERA", to.table = "Annotated")
FinalWorkflow(peak_db = peak_db, lib_db = lib_db)
}
}
