% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_metabolite_data.R
\name{format_MetabolomicData}
\alias{format_MetabolomicData}
\alias{format_MetabolomicData.pktable}
\alias{format_MetabolomicData.mass_all}
\alias{format_MetabolomicData.default}
\title{Formatting of metabolomic data for MetaboAnalystR.}
\usage{
format_MetabolomicData(mSetObj, Peak.list, Sample.df, Sample.data)

\method{format_MetabolomicData}{pktable}(mSetObj, Peak.list, Sample.df,
  Sample.data)

\method{format_MetabolomicData}{mass_all}(mSetObj, Peak.list, Sample.df,
  Sample.data)

\method{format_MetabolomicData}{default}(mSetObj, Peak.list, Sample.df,
  Sample.data)
}
\arguments{
\item{mSetObj}{NULL}

\item{Peak.list}{data frame containing combined ion mode peaklist with ion mode duplicates removed.}

\item{Sample.df}{data frame with class info as columns.  Must contain a separate row entry for each unique sex/class combination. Must contain the columns 'Sex','Class','n','Endogenous'.}

\item{Sample.data}{data frame with phenotype data as columns and a row for each study sample.  First column must be a unique sample identifier with the header 'CT-ID'.  Phenotype columns may vary, but must include two columns called 'Plate Number' and 'Plate Position' for determining run order.}
}
\value{
mSetObj
}
\description{
This function initializes objects that will hold the metabolite data, formats peak intensity data into one of the formats acceptable by MetaboAnalystR, and sets the metabolite data object.
}
