% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_Output-Modules.R
\name{FormatForMetaboAnalystR}
\alias{FormatForMetaboAnalystR}
\title{Formatting of metabolite data for MetaboAnalystR.}
\usage{
FormatForMetaboAnalystR(Peak.list, Sample.df, Sample.data,
  data.type = "pktable", anal.type = "stat", paired = FALSE)
}
\arguments{
\item{Peak.list}{data frame containing combined ion mode peaklist with ion mode duplicates removed.}

\item{Sample.df}{data frame with class info as columns.  Must contain a separate row entry for each unique sex/class combination. Must contain the columns 'Sex','Class','n','Endogenous'.}

\item{Sample.data}{ata frame with phenotype data as columns and a row for each study sample.  First column must be a unique sample identifier with the header 'CT-ID'.  Phenotype columns may vary, but must include two columns called 'Plate Number' and 'Plate Position' for determining run order.}

\item{data.type}{NULL}

\item{anal.type}{NULL}

\item{paired}{NULL}
}
\value{
mSetObj
}
\description{
This function initializes objects that will hold the metabolite data, formats peak intensity data into one of the formats acceptable by MetaboAnalystR, and sets the metabolite data object.
}
