% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_Output-Modules.R
\name{FormatForSIMCA}
\alias{FormatForSIMCA}
\title{Format Peaklists for input into SIMCA}
\usage{
FormatForSIMCA(from.table, to.csv, peak.db, db.dir)
}
\arguments{
\item{from.table}{from which table should LUMA read the Peaklist}

\item{to.csv}{to what filename (excluding .csv extension) should LUMA save
the formatted Peaklist}

\item{peak.db}{what database contains the Peaklists to be combined. Default
is \code{"Peaklist_db"}.}

\item{db.dir}{directory containing the database. Default is \code{"db"}.}
}
\description{
Formats Peaklists processed by LUMA and exports a csv file for
  direct import into the multivariate statistical software SIMCA. See
  \code{format_simca()} for more details. For examples, see
  \code{InitWorkflow()} and vignettes.
}
