% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-functions.R
\name{gen_filebase}
\alias{gen_filebase}
\title{Generate Filebase}
\usage{
gen_filebase(mzdatafiles, BLANK, ion.id, ion.mode)
}
\arguments{
\item{mzdatafiles}{character vector containing data files to process and store results in databases}

\item{BLANK}{a logical indicating whether blanks are being evaluated}

\item{ion.id}{character vector of length 2 specifying identifier in filename designating positive or negative ionization mode.  Positive identifier must come first.}

\item{ion.mode}{a character string defining the ionization mode.  Must be either 'Positive' or 'Negative'}
}
\value{
character
}
\description{
Generates filebase for reading and writing to databases
}
