% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-functions.R
\name{get_features}
\alias{get_features}
\title{Retrieves features from Peak.list in a database}
\usage{
get_features(mytbl, peak.db, asdf)
}
\arguments{
\item{mytbl}{character name of table in database to return}

\item{peak.db}{Formal class SQLiteConnection}

\item{asdf}{logical indicating whether to return a data frame instead of a tibble. Default is FALSE}
}
\value{
tbl alternatively a data frame
}
\description{
Returns mz/rt features from Peak.list stored in SQLite database
}
\examples{
library(LUMA)
if(require(RSQLite, quietly = TRUE)) {
file <- system.file("extdata","Sample_Data.csv", package =  "LUMA")
sample_data <- read.table(file, header = TRUE, sep = ",")
mzdatafiles <- sample_data$CT.ID
file.base <- gen_filebase(mzdatafiles = mzdatafiles, BLANK = FALSE, IonMode =
"Positive", ion.id = c("Pos","Neg")) #Returns "Peaklist_Pos"
peak_db <- connect_peakdb(file.base = file.base, mem = TRUE)
dbIsValid(peak_db) #Database is valid
dbListTables(peak_db) #But no tables yet
mydf <- Peaklist_Pos$From_CAMERA
write_tbl(mydf = mydf, myname = "From_CAMERA", peak.db = peak_db)
test_tbl <- LUMA:::get_features(mytbl = "From_CAMERA", peak.db = peak_db)
test_tbl #Returns a tibble

test_df <- LUMA:::get_features(mytbl = "From_CAMERA", peak.db = peak_db, asdf = TRUE)
test_df #Returns a data frame

dbDisconnect(peak_db)
}
}
