% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-data-functions.R
\name{LUMA_getPeaklist}
\alias{LUMA_getPeaklist}
\title{Generates Peak.list from CAMERA}
\usage{
LUMA_getPeaklist(object, convert.rt)
}
\arguments{
\item{object}{an xcmsSet or xsAnnotate object}

\item{convert.rt}{logical indicating whether to convert retention times to min}
}
\value{
data frame of peak intensities across all samples with metadata columns from XCMS and CAMERA
}
\description{
Generates Peak.list from xsAnnotate object generated from CAMERA, with the option to convert retention times to min
}
