% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize_Peaklists-Modules.R
\name{NormalizePeaklists}
\alias{NormalizePeaklists}
\title{Normalize Peaklists from different ion modes}
\usage{
NormalizePeaklists(from.table, to.table, peak.db, db.dir)
}
\arguments{
\item{from.table}{from which table should LUMA read the Peaklist}

\item{to.table}{to which table should LUMA save the modified Peaklist}

\item{peak.db}{what database contains the Peaklists to be combined.
Default is 'Peaklist_db'}

\item{db.dir}{directory containing the database.
Default is 'db'}
}
\description{
Normalizes Peaklists from positive and negative ionization modes separately.
See replace_zeros_normalize for more details.
}
