% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Munge_Data-Modules.R
\name{ParseCAMERA}
\alias{ParseCAMERA}
\title{Parses CAMERA annotations}
\usage{
ParseCAMERA(from.table, to.table, CAMERA.obj)
}
\arguments{
\item{from.table}{from which table should LUMA pull the Peaklist.}

\item{to.table}{to which should LUMA save the modified Peaklist.}

\item{CAMERA.obj}{\code{xsAnnotate} object used to plot metabolite groups.}
}
\description{
Parses the three annotation columns generated by CAMERA to
  remove redundant and duplicative ion adduct annotations. See
  \code{parse_pos_results(), parse_neg_results(), plot_metgroup()} for more
  details. For examples, see \code{InitWorkflow()} and vignettes.
}
