% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Munge_Data-Modules.R
\name{ParseCAMERA}
\alias{ParseCAMERA}
\title{Parses CAMERA annotations}
\usage{
ParseCAMERA(from.table, to.table, CAMERA.obj)
}
\arguments{
\item{from.table}{from which table should LUMA pull the Peak.list}

\item{to.table}{to which should LUMA save the modified Peak.list}

\item{CAMERA.obj}{which CAMERA object to use to plot metabolite groups.
Default is to read from saved R objects by call to InitWorkflow}
}
\description{
Parses the three annotation columns generated by CAMERA to remove redundant and duplicative ion adduct annotations
See parse_pos_results, parse_neg_results and plot_metgroup for more details.
}
\examples{
\dontrun{
library(LUMA)
db.dir <- system.file('extdata/', package = "LUMA")
InitWorkflow(db.dir = db.dir)
ParseCAMERA(from.table = "From CAMERA", to.table = "output_parsed")
}
}
