% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Munge_Data-Modules.R
\name{ParseCAMERA}
\alias{ParseCAMERA}
\title{Parses CAMERA annotations}
\usage{
ParseCAMERA(from.table, to.table, CAMERA.obj)
}
\arguments{
\item{from.table}{from which table should LUMA pull the Peaklist.}

\item{to.table}{to which should LUMA save the modified Peaklist.}

\item{CAMERA.obj}{\code{xsAnnotate} object used to plot metabolite groups.}
}
\description{
Parses the three annotation columns generated by CAMERA to
  remove redundant and duplicative ion adduct annotations. See
  \code{parse_pos_results, parse_neg_results, plot_metgroup} for more
  details. For working examples, see \code{InitWorkflow, AnnotatePeaklist,
  CombineFeatures, CombinePeaklists, CullBackground, CullMF, CullCV,
  CullVoidVolume, FormatForMetaboAnalystR, FormatForSIMCA,
  NormalizePeaklists, ParseCAMERA, SimplyPeaklists, FinalWorkflow}.
}
\examples{
\dontrun{
library(LUMA)
if(require(lcmsfishdata, quietly = TRUE)) {
  db.dir <- system.file('extdata/', package = "lcmsfishdata")
  InitWorkflow(db.dir = db.dir)
  AnnotatePeaklist(from.table = "From_CAMERA", to.table = "Annotated")
  ParseCAMERA(from.table = "Annotated", to.table = "output_parsed", CAMERA.obj = anposGa)
  FinalWorkflow(peak_db = peak_db, lib_db = lib_db)
  }
}
}
