% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting-functions.R
\name{plot_ionduplicate}
\alias{plot_ionduplicate}
\title{Loop-based EIC Plotter for Ion Mode Duplicates}
\usage{
plot_ionduplicate(anposGa, xpos.cor, annegGa, xneg.cor, Peak.list,
  gen.plots, file.base, QC.id, ...)
}
\arguments{
\item{anposGa}{xsannotate object with annotated isotopes and ion adducts and fragments for positive mode.}

\item{xpos.cor}{xcmsSet object shoud have grouping, retention time correction and fillPeaks applied.  Default is to look for this in anposGa}

\item{annegGa}{xsannotate object with annotated isotopes and ion adducts and fragments for negative mode.}

\item{xneg.cor}{xcmsSet object shoud have grouping, retention time correction and fillPeaks applied.  Default is to look for this in annegGa}

\item{Peak.list}{a data frame containing combined ion mode peaklist with Duplicate IDs.  Alternatively can be retrieved from databases.  Default is NULL}

\item{gen.plots}{a logical indicating whether to create plots for metabolite groups.  Default is FALSE}

\item{file.base}{character string used to name graphical output. Default is 'EIC_plots'}

\item{QC.id}{character identifier for pooled QC samples. Default is 'Pooled_QC'}

\item{...}{parameters to be passed to database functions}
}
\value{
NULL testing
}
\description{
Plots EICs and Pseudospectra from (parsed) metabolite groups using plotEICs-methods and plotPsSpectrum-methods from CAMERA. Also plots correlation matrices and clustered dendrograms for each (parsed) metabolite group.
}
