% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-functions.R
\name{read_tbl}
\alias{read_tbl}
\title{Reads Peak.list from database}
\usage{
read_tbl(mytable, peak.db, asdf)
}
\arguments{
\item{mytable}{character name of table in database to return}

\item{peak.db}{Formal class SQLiteConnection}

\item{asdf}{logical indicating whether to return a data frame instead of a tibble. Default is FALSE}
}
\value{
tbl alternatively a data frame
}
\description{
Extracts Peak.list from an SQLite database as a tibble.
Alternatively, load the Peak.list into memory as a data frame
}
\examples{
library(LUMA)
if(require(RSQLite, quietly = TRUE)) {
file <- system.file("extdata","Sample_Data.csv", package =  "LUMA")
sample_data <- read.table(file, header = TRUE, sep = ",")
mzdatafiles <- sample_data$CT.ID
file.base <- gen_filebase(mzdatafiles = mzdatafiles, BLANK = FALSE, IonMode =
"Positive", ion.id = c("Pos","Neg")) #Returns "Peaklist_Pos"
peak_db <- connect_peakdb(file.base = file.base, mem = TRUE)
dbIsValid(peak_db) #Database is valid
dbListTables(peak_db) #But no tables yet
mydf <- Peaklist_Pos$From_CAMERA
write_tbl(mydf = mydf, myname = "From_CAMERA", peak.db = peak_db)
test <- read_tbl(mytable = "From_CAMERA", peak.db = peak_db)
identical(test,mydf) ## Objects are identical of class "tbl"
test2 <- read_tbl(mytable = "From_CAMERA", peak.db = peak_db, asdf = TRUE)
!identical(test2,mydf) ## test2 is a data.frame
}
}
