% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mummichog-functions.R
\name{run_mummichog}
\alias{run_mummichog}
\title{Runs \emph{mummichog}}
\usage{
run_mummichog(Sample.df, Peak.list, ion.mode, stat.method, search.par,
  instrumentOpt, msModeOpt, pvalCutoff, lib, enrichOpt, pvalOpt, permNum)
}
\arguments{
\item{Sample.df}{a data frame with class info as columns.  Must contain a separate row entry for each unique sex/class combination. Must contain the columns 'Sex','Class','n','Endogenous'.}

\item{Peak.list}{a table of class 'tbl_df',tbl' or 'data.frame' with variables as columns.  Should contain all output columns from XCMS and CAMERA.}

\item{ion.mode}{a character string defining the ionization mode.  Must be either 'Positive' or 'Negative'}

\item{stat.method}{statistical method to use. Can be "TTest" or "ANOVA"}

\item{search.par}{a single-row data frame with 11 variables containing user-defined search parameters. Must contain the columns 'ppm','rt','Voidrt','Corr.stat.pos','Corr.stat.neg','CV','Minfrac','Endogenous','Solvent','gen.plots','keep.singletons'.}

\item{instrumentOpt}{Define the mass-spec instrument used to perform untargeted metabolomics}

\item{msModeOpt}{Define the mass-spec mode of the instrument used to perform untargeted metabolomics}

\item{pvalCutoff}{Numeric, specify the p-value cutoff to define significant m/z features from reference m/z features}

\item{lib}{Input the name of the organism library, default is hsa}

\item{enrichOpt}{Input the method to perform enrichment analysis}

\item{pvalOpt}{Input the method to calculate p-values}

\item{permNum}{Numeric, the number of permutations to perform}
}
\value{
NULL testing
}
\description{
Runs \emph{mummichog} within the R environment.  Basically a wrapper script built from the \bold{"MS Peaks to Pathways"} vignette from MetaboAnalystR.
}
