% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dialog-functions.R
\name{ScriptInfo_dlg}
\alias{ScriptInfo_dlg}
\title{Script Info Dialog Box}
\usage{
ScriptInfo_dlg(multiple = FALSE)
}
\arguments{
\item{multiple}{should dialog box with multiple fields be used; default if
FALSE}
}
\value{
named list
}
\description{
Creates dialog box for user to set basic workflow (used to be
  called script) information
}
\examples{
\dontrun{
library(LUMA)
if(require(lcmsfishdata, quietly = TRUE)) {
  (mypath <- system.file("extdata", package = "LUMA"))
  (test <- ScriptInfo_dlg(multiple = TRUE))
  }
}
}
