% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-functions.R
\name{search_IonDup}
\alias{search_IonDup}
\title{Searches for Ion mode duplicates}
\usage{
search_IonDup(object, Peak.list.pos, Peak.list.neg, search.par)
}
\arguments{
\item{object}{used for method dispatch. Can be any object. Class must be "mz" or "monoMass"}

\item{Peak.list.pos}{Positive ionization mode data table}

\item{Peak.list.neg}{Negative ionization mode data table}

\item{search.par}{a single-row data frame with 11 variables containing user-defined search parameters. Must contain the columns 'ppm','rt','Voidrt','Corr.stat.pos','Corr.stat.neg','CV','Minfrac','Endogenous','Solvent','gen.plots','keep.singletons'.}
}
\value{
data frame containing the original table with added columns 'Name','MS.ID','Formula','Annotated.adduct' and any additional info columns from Annotated.library
}
\description{
Searches data tables from both ionization modes for duplicate entries
}
