% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-functions.R
\name{search_solv}
\alias{search_solv}
\alias{search_solv.mz}
\alias{search_solv.monoMass}
\title{Searches Peak.list for background components}
\usage{
search_solv(object, Peak.list, Solv.list, Sample.df, search.par, ion.id,
  QC.id, MB.id, db.id, ion.modes, lib_db)

\method{search_solv}{mz}(object, Peak.list, Solv.list, Sample.df,
  search.par, ion.id, QC.id, MB.id, db.id, ion.modes, lib_db)

\method{search_solv}{monoMass}(object, Peak.list, Solv.list, Sample.df,
  search.par, ion.id, QC.id, MB.id, db.id, ion.modes, lib_db)
}
\arguments{
\item{object}{used for method dispatch. Can be any object. See usage for details}

\item{Peak.list}{data table containing sample data}

\item{Solv.list}{data table containing blank data}

\item{Sample.df}{a data frame with class info as columns.  Must contain a separate row entry for each unique sex/class combination. Must contain the columns 'Sex','Class','n','Endogenous'.}

\item{search.par}{a single-row data frame with 11 variables containing user-defined search parameters. Must contain the columns 'ppm','rt','Voidrt','Corr.stat.pos','Corr.stat.neg','CV','Minfrac','Endogenous','Solvent','gen.plots','keep.singletons'.}

\item{ion.id}{character vector specifying identifier in filename designating positive or negative ionization mode or both.  Positive identifier must come first. Default is c('Pos','Neg')}

\item{QC.id}{character vector specifying identifier in filename designating a Pooled QC sample.  Only the first value will be used.  Default is 'Pooled_QC_'}

\item{MB.id}{character vector specifying identifier in filename designating a Method Blank.  Only the first value will be used. Default is '^MB'}

\item{db.id}{character vector specifying identifiers in database names designating sample \[1\] and blank \[2\] databases.  Default is c('Peaklist','Blanks')}

\item{ion.modes}{a character vector defining the ionization mode.  Must be either 'Positive', 'Negative' or both.}

\item{lib_db}{RSQLite connection}
}
\value{
nested list a list for each ionization mode, each containing a list of two dataframes: the first contains the intensity matrix for the peaklist with solvent peaks removed, the second contains the intensity matrix for the solvent peaks
}
\description{
Searches Peak.list containing all samples for components that are present in the process blanks
}
