% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Combine_Peaklists-Modules.R
\name{SimplyPeaklists}
\alias{SimplyPeaklists}
\title{Combine Peaklists simply}
\usage{
SimplyPeaklists(from.tables, to.table, peak.db, db.dir)
}
\arguments{
\item{from.tables}{character vector of table names to draw from databases to
be combined simply.}

\item{to.table}{to which table should LUMA save the modified Peaklist}

\item{peak.db}{what database contains the Peaklists to be combined simply.
Default is \code{"Peaklist_db"}}

\item{db.dir}{directory containing the database. Default is 'db'}
}
\value{
the file \code{"Peaklist_Combined.csv"} is written to the working
  directory
}
\description{
Combines Peaklists from positive and negative ionization modes
  without removing ion mode duplicates. For examples, see
  \code{InitWorkflow()} and vignettes.
}
\examples{
\dontrun{
library(LUMA)
if(require(lcmsfishdata, quietly = TRUE)) {
  db.dir <- system.file("extdata", package = "lcmsfishdata")
  InitWorkflow(db.dir = db.dir)
  SimplyPeaklists(from.tables = c("Peaklist_Pos_Solvent Peaks Removed",
  "Peaklist_Neg_Solvent Peaks Removed"),
  to.table = "Peaklist_Combined",
  peak.db = peak_db, db.dir = db.dir)
  FinalWorkflow(peak_db = peak_db, lib_db = lib_db)
  }
}
}
