% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimming-functions.R
\name{trim_rt}
\alias{trim_rt}
\title{Trims by retention time}
\usage{
trim_rt(Peak.list, ...)
}
\arguments{
\item{Peak.list}{data frame. Must have Correlation.stat column.  Should contain output columns from XCMS and CAMERA, and additional columns from IHL.search, Calc.MinFrac, Calc.corr.stat and EIC.plotter functions.}

\item{...}{arguments to pass to other functions}
}
\description{
Removes features with retention times smaller than the user specified threshold.
}
