% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-functions.R
\name{wrap_camera}
\alias{wrap_camera}
\title{Wraps CAMERA}
\usage{
wrap_camera(xcms.obj, CAMERA.par, IonMode)
}
\arguments{
\item{xcms.obj}{an \code{xcmsSet} object that has had peak picking, retention time
alignment, peak grouping, and imputing missing values performed.}

\item{CAMERA.par}{a single-row data frame with 9 variables containing CAMERA
parameters. Must include columns \code{c("perfwhm","sigma","minfrac","mzabs",
"maxiso","corval_eic","corval_exp","pval","mzabs.1")}.}

\item{IonMode}{a character string defining the ionization mode.  Must be one
of \code{c("Positive","Negative")}.}
}
\value{
two grouped \code{xsannotate} objects \code{mz1setpos,anposGa}
  without and with annotated isotopes and ion adducts and fragments,
  respectively.
}
\description{
Run a simple CAMERA workflow with user defined input parameters
  and return xsAnnotate objects
}
\examples{
library(LUMA)
if(require(lcmsfishdata, quietly = TRUE)) {
file <- system.file("extdata","XCMS_objects_pos.Rdata", package =
"lcmsfishdata") # is case sensitive on Linux
load(file, envir = environment())
file2 <- system.file("extdata","Best_CAMERA_parameters_positive.csv", package
= "lcmsfishdata")  # is case sensitive on Linux
CAMERA.par <- read.table(file2, header = TRUE, sep = ",")

\dontrun{

#Runs CAMERA. This requires access to raw datafiles and won't work with
#lcmsfishdata. Better to use your own data here.
test <- wrap_camera(xcms.obj = xset4, CAMERA.par = CAMERA.par, IonMode = "Positive")
}

}
}
