% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database-functions.R
\name{write_tbl}
\alias{write_tbl}
\title{Writes table to database}
\usage{
write_tbl(mytbl, peak.db, myname)
}
\arguments{
\item{mytbl}{tbl or dataframe to write}

\item{peak.db}{Formal class SQLiteConnection}

\item{myname}{character what should the table be called}
}
\value{
a tbl object in the remote source
}
\description{
Writes tbl or dataframe to an RSQLite database
}
